Allow peeking over more than one token
This is needed for more complex syntax such as python's
expr if condition else expr

Also remove the hack that was present for trying to support JS literal
regexp syntax at hte start of a statement. It did not work and it
messed with the internal state of the parser in unsafe ways.

...