for..in now works with ES6 collections
It now works with arrays, strings, any object that implements the ES 6
iterable protocol and arbitrary objects via Object.keys(). Its only
overhead for arrays is a single function call that does nothing.

The implementation for other types is not efficient, but that cant be
helped until we drop support of ES 5 only runtimes.

Also do the same for the enumerate() and reversed() builtin functions