Make truthyness in RS the same as in Python
This has a small performance impact as every conditional now involves a
function call to _$rapyd$_bool. However, it can be easily optimized in
hot code paths by using the JS() literal. For example,

if JS('x.length > 0'):
  ...

Since the difference in truthyness is pretty surprising to new RS
developers and easy to accidentally forget for even experienced
developers, I consider the tradeoff to be worthwhile.