Remove the few remaining optional baselib components
Now the only optional component is support for yield on ES 5, which
is only included if actually used. All other parts of the baselib are
included always.

This was necessary because trying to support function annotations for
optional baselib functions was too much work, and in any case the remaining
optional functions were very few compared to the rest of the baselib.

Allows for a great simplification of the compiler bootstrapping code.